+++ website/src/components/ConnectingBanner.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ConnectingBanner = () => {
+    return (
+        <div className="m-5 p-5 text-center">
+            <h2>Trying to Connect...</h2>
+        </div>
+    );
+};
+
+export default ConnectingBanner;
