+++ website/src/services/model.js
@@ -1,0 +1,20 @@
+// export type ThemeType = "default" | "light" | "dark";
+
+// export interface SettingsType {
+//   theme: ThemeType;
+//   fontSize: number;
+//   saveLocal: boolean;
+// }
+
+// export interface DataValueType {
+//   lastUpdate: number;
+//   value: string;
+//   users: number;
+// }
+
+// export interface SuccessDataResponseType {
+//   status: string;
+//   data?: DataValueType;
+//   room_id?: string;
+//   users?: number;
+// }
